Debugged
Debugged: Provided quotation marks to cater for file or folder names with special characters such  as white space and (). Without this quotation mark, folders had to be created with all white spaces replaced with _ before ffmpeg considered it to be a valid string.		   Another flaw was that folder names with () in them were considered invalid strings, thus, were not download.